# Anonymous-Proxy-Configurator/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: MultiProxy Anonymous Router API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /system/config:
    get:
      operationId: getConfig
      tags: [system]
      summary: Get application configuration
      responses:
        "200":
          description: Config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppConfig"
    put:
      operationId: updateConfig
      tags: [system]
      summary: Update application configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppConfig"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppConfig"
...
  /system/rotation/start:
    post:
      operationId: startProxyRotation
      tags: [system]
      summary: Start automatic proxy rotation
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                intervalSeconds:
                  type: integer
                  description: Rotation interval in seconds
      responses:
        "200":
          description: Rotation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceResult"
...
  /proxies:
    get:
      operationId: listProxies
      tags: [proxies]
      summary: List all configured proxies
      responses:
        "200":
          description: List of proxies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Proxy"
    post:
      operationId: addProxy
      tags: [proxies]
      summary: Add a new proxy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProxy"
      responses:
        "201":
          description: Proxy added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Proxy"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /proxies/bulk-upload:
    post:
      operationId: bulkUploadProxies
      tags: [proxies]
      summary: Bulk import proxies from a text list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: "Newline-separated list of proxies (format: type:host:port or host:port)"
                defaultType:
                  type: string
                  enum: [socks5, socks4, http, https]
                  default: socks5
              required: [text]
      responses:
        "200":
          description: Bulk upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkUploadResult"
...
  /proxies/{id}:
    delete:
      operationId: deleteProxy
      tags: [proxies]
      summary: Delete a proxy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proxy deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /proxies/{id}/validate:
    post:
      operationId: validateProxy
      tags: [proxies]
      summary: Validate if a proxy is alive
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Proxy validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxyValidationResult"
...
components:
  schemas:
...
    AppConfig:
      type: object
      properties:
        torPort:
          type: integer
          default: 9050
        torControlPort:
          type: integer
          default: 9051
        rotationIntervalSeconds:
          type: integer
          default: 180
        autoRemoveDeadProxies:
          type: boolean
          default: false
        logFilePath:
          type: string
        proxychainsConfigPath:
          type: string
        maxRetries:
          type: integer
          default: 3
      required: [torPort, torControlPort, rotationIntervalSeconds, autoRemoveDeadProxies]
...
    ServiceResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        status:
          type: string
      required: [success, message, status]
...
    BulkUploadResult:
      type: object
      properties:
        added:
          type: integer
        skipped:
          type: integer
        errors:
          type: integer
        total:
          type: integer
        message:
          type: string
      required: [added, skipped, errors, total, message]
    Proxy:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [socks5, socks4, http, https]
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        password:
          type: string
        enabled:
          type: boolean
        status:
          type: string
          enum: [unknown, alive, dead, checking]
        latencyMs:
          type: integer
        rating:
          type: string
          enum: [fast, medium, slow, unknown]
        lastChecked:
          type: string
        createdAt:
          type: string
      required:
        - id
        - type
        - host
        - port
        - enabled
        - status
        - rating
        - createdAt
    CreateProxy:
      type: object
      properties:
        type:
          type: string
          enum: [socks5, socks4, http, https]
        host:
          type: string
        port:
          type: integer
        username:
          type: string
        password:
          type: string
        enabled:
          type: boolean
      required: [type, host, port]
    ProxyValidationResult:
      type: object
      properties:
        proxyId:
          type: integer
        host:
          type: string
        port:
          type: integer
        alive:
          type: boolean
        latencyMs:
          type: integer
        rating:
          type: string
          enum: [fast, medium, slow, unknown]
        message:
          type: string
      required: [proxyId, host, port, alive, rating]
...
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
      required: [error]
...