# Anonymous-Proxy-Configurator/lib/api-zod/src/generated/types/proxyValidationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MultiProxy Anonymous Router API
 * OpenAPI spec version: 0.2.0
 */
import type { ProxyValidationResultRating } from "./proxyValidationResultRating";

export interface ProxyValidationResult {
  proxyId: number;
  host: string;
  port: number;
  alive: boolean;
  latencyMs?: number;
  rating: ProxyValidationResultRating;
  message?: string;
}