# Anonymous-Proxy-Configurator/lib/api-zod/src/generated/types/serviceResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MultiProxy Anonymous Router API
 * OpenAPI spec version: 0.2.0
 */

export interface ServiceResult {
  success: boolean;
  message: string;
  status: string;
}